[feature] migrate plotting functions from v1_depth_map.
